+++ src/serena/util/yaml.py
@@ -4,7 +4,7 @@ from collections.abc import Sequence
 from enum import Enum
 from typing import Any
 
-from ruamel.yaml import YAML, CommentToken, StreamMark
+from ruamel.yaml import YAML, CommentedSeq, CommentToken, StreamMark
 from ruamel.yaml.comments import CommentedMap
 
 from serena.constants import SERENA_FILE_ENCODING
@@ -122,6 +122,20 @@ def normalise_yaml_comments(commented_map: CommentedMap, comment_normalisation: 
                 first_token.value = first_token.value[1:]
         return token_list
 
+    def remove_nested_comments() -> None:
+        """
+        Removes nested comments, particularly of sequences, which incorrectly capture comments
+        that are actually intended for top-level keys.
+        """
+        for key in commented_map.keys():
+            entry = commented_map[key]
+            if isinstance(entry, CommentedSeq):
+                items = entry.ca.items
+                if isinstance(items, dict):
+                    items_keys = list(items.keys())
+                    for i in items_keys:
+                        items[i] = [None] * 4
+
     match comment_normalisation:
         case YamlCommentNormalisation.NONE:
             pass
@@ -176,6 +190,9 @@ def normalise_yaml_comments(commented_map: CommentedMap, comment_normalisation: 
                                     preceding_comment[ITEM_COMMENT_INDEX_BEFORE] = token_list
                                     current_comment[ITEM_COMMENT_INDEX_BEFORE] = None
                     preceding_comment = current_comment
+
+            # remove nested comments, as we assume that only top-level keys are supposed to be commented
+            remove_nested_comments()
         case _:
             raise ValueError(f"Unhandled comment normalisation: {comment_normalisation}")
 
@@ -204,7 +221,7 @@ def yaml_comment_entry_is_empty(comment_entry: Any) -> bool:
         return False
 
 
-def transfer_missing_yaml_comments_by_index(
+def transfer_yaml_comments_by_index(
     source: CommentedMap, target: CommentedMap, indices: list[int], forced_update_keys: Sequence[str] = (), force_update_all: bool = False
 ) -> None:
     """
@@ -231,7 +248,7 @@ def transfer_missing_yaml_comments_by_index(
                     target_comment[index] = source_comment[index]
 
 
-def transfer_missing_yaml_comments(
+def transfer_yaml_comments(
     source: CommentedMap,
     target: CommentedMap,
     comment_normalisation: YamlCommentNormalisation,
@@ -251,7 +268,7 @@ def transfer_missing_yaml_comments(
         case YamlCommentNormalisation.NONE:
             pass
         case YamlCommentNormalisation.LEADING | YamlCommentNormalisation.LEADING_WITH_CONVERSION_FROM_TRAILING:
-            transfer_missing_yaml_comments_by_index(
+            transfer_yaml_comments_by_index(
                 source, target, [ITEM_COMMENT_INDEX_BEFORE], forced_update_keys=forced_update_keys, force_update_all=force_update_all
             )
         case _: